builder: adjust the flake to be usable in Nix sandbox.
The original flake only works when Nix is running with no sandbox (i.e., in a
container with no privileges). The packages specific for development (like meson
or ccls) are also removed.

The flake can also built the FFmpeg binary with the drawvg filter.

Now, emscripten is installed from nixpkgs, and upgraded to 4.0.11. An overlay is
used to prevent installing the default JRE (which has a lot of unneeded deps.).